產生報表前後的word程序的清除

+++ WPFSTD105Base/ViewModel/OfficeTypeSettingVM.cs
@@ -50,7 +50,9 @@ namespace WPFSTD105
         {
 
         }
-        //1201 CYH
+        /// <summary>
+        /// 報表下載等待時間的提示
+        /// </summary>
         private static SplashScreenManager manager = SplashScreenManager.Create(() => new ProcessingScreenWindow(), new DXSplashScreenViewModel { });
 
         public static void ActivateLoading()
@@ -168,6 +170,14 @@ namespace WPFSTD105
             {
                 return new RelayParameterizedCommand(el =>
                 {
+                    string current_time = DateTime.Now.ToString("yyyyMMdd");
+
+                    //刪除目前已有WINWORD程序, 以避免發生不可預期的錯誤
+                    try
+                    {
+                        DeleteWordFileAfterDelay(100, "", "");
+                    }
+                    catch (Exception ex) { Console.WriteLine(ex.ToString()); }
                     //ExcelCutService execl = new ExcelCutService();
                     //execl.CreateFile($@"{Properties.SofSetting.Default.LoadPath}\{CommonViewModel.ProjectName}\切割明細表.xls", MaterialDataViews);
 
@@ -200,7 +210,7 @@ namespace WPFSTD105
                         FloatingMode.Popup);
                     }
 
-                    string current_time = DateTime.Now.ToString("yyyyMMdd");
+                    //string current_time = DateTime.Now.ToString("yyyyMMdd");
 
                     try 
                     { 
@@ -236,7 +246,8 @@ namespace WPFSTD105
                         FloatingMode.Popup);
                     }
 
-                    try 
+                    //刪除已產生的報表WORD檔&目前所有的WINWORD程序
+                    try
                     {
                         DeleteWordFileAfterDelay(3000, $@"{Properties.SofSetting.Default.LoadPath}\{CommonViewModel.ProjectName}\切割明細表.docx", $@"{Properties.SofSetting.Default.LoadPath}\{CommonViewModel.ProjectName}\切割明細表_{current_time}.pdf");
                     }
@@ -285,6 +296,15 @@ namespace WPFSTD105
             {
                 return new RelayParameterizedCommand(el =>
                 {
+                    string current_time = DateTime.Now.ToString("yyyyMMdd");
+
+                    //刪除目前已有WINWORD程序, 以避免發生不可預期的錯誤
+                    try
+                    {
+                        DeleteWordFileAfterDelay(100, "", "");
+                    }
+                    catch (Exception ex) { Console.WriteLine(ex.ToString()); }
+
                     //ExcelBuyService execl = new ExcelBuyService();
                     //execl.CreateFile($@"{Properties.SofSetting.Default.LoadPath}\{CommonViewModel.ProjectName}\採購明細單.xls", MaterialDataViews);
 
@@ -317,7 +337,7 @@ namespace WPFSTD105
                         FloatingMode.Popup);
                     }
 
-                    string current_time = DateTime.Now.ToString("yyyyMMdd");
+                    //string current_time = DateTime.Now.ToString("yyyyMMdd");
 
                     try
                     {
@@ -353,12 +373,13 @@ namespace WPFSTD105
                         FloatingMode.Popup);
                     }
 
+                    //刪除已產生的報表WORD檔&目前所有的WINWORD程序
                     try
                     {
                         DeleteWordFileAfterDelay(3000, $@"{Properties.SofSetting.Default.LoadPath}\{CommonViewModel.ProjectName}\採購明細單.docx", $@"{Properties.SofSetting.Default.LoadPath}\{CommonViewModel.ProjectName}\採購明細單_{current_time}.pdf");
                     }
                     catch (Exception ex) { Console.WriteLine(ex.ToString()); }
-                    
+
 
                     ////以python封裝exe執行轉檔                    
                     //try
@@ -429,33 +450,35 @@ namespace WPFSTD105
             //底線的寫法代表使用lambda時，如果不使用到傳入參數，就要打一個底線。
             System.Threading.Tasks.Task.Delay(Delay_ms).ContinueWith(_ => 
             {
-                if (File.Exists(PdfFilePath))
+                if (WordFilePath != "")
                 {
-                    //刪除word檔案
-                    try 
-                    { 
-                        File.Delete(WordFilePath);
-                    }
-                    catch (Exception ex) { Console.WriteLine(ex.ToString()); }
-
-                    //刪除存留的word程序
-                    try
+                    if (File.Exists(PdfFilePath))
                     {
-                        Process[] processes = System.Diagnostics.Process.GetProcessesByName("WINWORD");
-
-                        foreach (System.Diagnostics.Process process in processes)
+                        //刪除word檔案
+                        try
                         {
-                            if (process.MainWindowTitle == "")
-                            {
-                                process.Kill();
-
-                                System.Threading.Thread.Sleep(500);
-                            }
-                            GC.Collect();
+                            File.Delete(WordFilePath);
                         }
+                        catch (Exception ex) { Console.WriteLine(ex.ToString()); }
                     }
-                    catch (Exception ex) { Console.WriteLine(ex.ToString()); }
                 }
+                //刪除存留的word程序
+                try
+                {
+                    Process[] processes = System.Diagnostics.Process.GetProcessesByName("WINWORD");
+
+                    foreach (System.Diagnostics.Process process in processes)
+                    {
+                        if (process.MainWindowTitle == "")
+                        {
+                            process.Kill();
+
+                            System.Threading.Thread.Sleep(500);
+                        }
+                        GC.Collect();
+                    }
+                }
+                catch (Exception ex) { Console.WriteLine(ex.ToString()); }
             });  
         }
     }
